(refactoring) move workshop invitation management to WorkshopInvitationManagement module and resolve rubocop issues
